# openapi/audio-studio-v1.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "VORVN Audio Studio API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/asset-collections/{collection_id}/assets/upload": {
      "post": {
        "operationId": "uploadVentureAsset",
        "parameters": [
          {
            "in": "path",
            "name": "collection_id",
            "required": true,
            "schema": {
              "title": "Collection Id",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "x-filename",
            "required": false,
            "schema": {
              "default": "audio.mp3",
              "title": "X-Filename",
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Uploadventureasset",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Upload Venture Asset",
        "tags": [
          "uploads"
        ]
      }
    },
...
    "/api/v1/exports/{export_id}/download": {
      "get": {
        "operationId": "downloadExport",
        "parameters": [
          {
            "in": "path",
            "name": "export_id",
            "required": true,
            "schema": {
              "title": "Export Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Download Export",
        "tags": [
          "media"
        ]
      }
    },
    "/api/v1/generations/{generation_id}/download": {
      "get": {
        "operationId": "downloadGeneration",
        "parameters": [
          {
            "in": "path",
            "name": "generation_id",
            "required": true,
            "schema": {
              "title": "Generation Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Download Generation",
        "tags": [
          "media"
        ]
      }
    },
...
    "/api/v1/project-covers/upload": {
      "post": {
        "operationId": "uploadProjectCover",
        "parameters": [
          {
            "in": "header",
            "name": "x-filename",
            "required": false,
            "schema": {
              "default": "image.png",
              "title": "X-Filename",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Uploadprojectcover",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Upload Project Cover",
        "tags": [
          "uploads"
        ]
      }
    },
...
    "/api/v1/subtitles": {
      "get": {
        "operationId": "listSubtitles",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 40,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Listsubtitles",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Subtitles",
        "tags": [
          "subtitles"
        ]
      }
    },
    "/api/v1/subtitles/uploads": {
      "post": {
        "operationId": "uploadSubtitleSource",
        "parameters": [
          {
            "in": "header",
            "name": "x-filename",
            "required": false,
            "schema": {
              "default": "audio.mp3",
              "title": "X-Filename",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Uploadsubtitlesource",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Upload Subtitle Source",
        "tags": [
          "uploads"
        ]
      }
    },
    "/api/v1/subtitles/{transcript_id}": {
      "delete": {
        "operationId": "deleteSubtitle",
        "parameters": [
          {
            "in": "path",
            "name": "transcript_id",
            "required": true,
            "schema": {
              "title": "Transcript Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Deletesubtitle",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Subtitle",
        "tags": [
          "subtitles"
        ]
      },
      "get": {
        "operationId": "getSubtitle",
        "parameters": [
          {
            "in": "path",
            "name": "transcript_id",
            "required": true,
            "schema": {
              "title": "Transcript Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Getsubtitle",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Subtitle",
        "tags": [
          "subtitles"
        ]
      }
    },
    "/api/v1/subtitles/{transcript_id}/layouts/{profile}": {
      "get": {
        "description": "Derive a presentation from saved timings without another provider call.",
        "operationId": "getSubtitleLayout",
        "parameters": [
          {
            "in": "path",
            "name": "transcript_id",
            "required": true,
            "schema": {
              "title": "Transcript Id",
              "type": "integer"
            }
          },
          {
            "in": "path",
            "name": "profile",
            "required": true,
            "schema": {
              "enum": [
                "standard",
                "short",
                "words"
              ],
              "title": "Profile",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Getsubtitlelayout",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Subtitle Layout",
        "tags": [
          "subtitles"
        ]
      }
    },
...
    "/api/v1/venture-logos/upload": {
      "post": {
        "operationId": "uploadVentureLogo",
        "parameters": [
          {
            "in": "header",
            "name": "x-filename",
            "required": false,
            "schema": {
              "default": "image.png",
              "title": "X-Filename",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Uploadventurelogo",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Upload Venture Logo",
        "tags": [
          "uploads"
        ]
      }
    },
...
    "/api/v1/voice-images/upload": {
      "post": {
        "operationId": "uploadVoiceImage",
        "parameters": [
          {
            "in": "header",
            "name": "x-filename",
            "required": false,
            "schema": {
              "default": "image.png",
              "title": "X-Filename",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Uploadvoiceimage",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Upload Voice Image",
        "tags": [
          "uploads"
        ]
      }
    },
...
    "/api/v1/voice-references/upload": {
      "post": {
        "operationId": "uploadVoiceReference",
        "parameters": [
          {
            "in": "header",
            "name": "x-filename",
            "required": false,
            "schema": {
              "default": "reference.wav",
              "title": "X-Filename",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Uploadvoicereference",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Upload Voice Reference",
        "tags": [
          "uploads"
        ]
      }
    },
...
  }
}